+++ components/sidebar/post-list-bar.tsx
@@ -6,11 +6,13 @@ export default async function PostListBar() {
   const postsTree = await getPostsTree();
 
   return (
-    <div className="bg-[rgba(255,255,255,0.5)] p-5 rounded-lg font-bold sticky top-2">
-      <h2 className="border-b-2 border-solid border-slate-300 mb-2">
+    <div className="bg-[rgba(255,255,255,0.5)] p-5 rounded-lg sticky top-2">
+      <h2 className="border-b-2 border-solid border-slate-300 mb-2 pb-2 font-semibold">
         文章目录
       </h2>
-      <PostTreeNode node={postsTree} />
+      <div className="leading-loose">
+        <PostTreeNode node={postsTree} />
+      </div> 
     </div>
   );
 }